requirements.txt
flask>=3.0.0
requests>=2.31.0
python-dotenv>=1.1.1
rich>=13.7.0
Jinja2>=3.1.3